Incorporate TokenizerType.swift to Mustard.swift and update tests.
Segmentation fault 11 were prevented by combining TokenizerType and
Mustard source files (╯°□°)╯︵ ┻━┻
