+++ Playgrounds/Source/Tokenize by chunking groups of characters.playground/Contents.swift
@@ -31,7 +31,7 @@ class ChunkingTokenizer: TokenizerType {
 }
 
 // create a tokenzier that chunks text into groups of 10 *scalars*
-let chunker = ChunkingTokenizer(chunkSize: 10).anyTokenizer
+let chunker = ChunkingTokenizer(chunkSize: 10)
 
 let sequence = "ACAAGATGCCATTGTCCCCCGGCCTCCTGCTGCTGCTGCTCTCCGGGGCCACGGCCACCGCTGCCCTGCCCCGGAGGGTGGCCCCACCGGCCGAGACAGCGAGCATATGCAGGAAGCGGCAGGAATAAGGAAAAGCAGCCTCCTGACTTTCCTCGCTTGGTGGTTTGAGTGGACCTCCCAGGCCAGTGCCGGGCCCCTCATAGGAGAGGAAGCTCGGGAGGTGGCCAGGCGGCAGGAAGGCGCACCCCCCCAGCAATCCGCGCGCCGGGACAGAATGCCCTGCAGGAACTTCTTCTGGAAGACCTTCTCCTCCTGCAAATAAAACCTCACCCATGAATGCTCACGCAAGTTTAATTACAGACCTGAA"
