+++ Tests/EmojiTokenTests.swift
@@ -80,7 +80,7 @@ class EmojiTokenTests: XCTestCase {
         // -> 7 (4 base, combied with 3 zero-width joiners \u{200D})
         
         let sample = "baby:👶 baby:👶🏿 flag:🇳🇿 flag:🏳️‍🌈 family:👩‍👩‍👦‍👦"
-        let tokens = sample.tokens(matchedWith: EmojiTokenizer.defaultTokenzier)
+        let tokens = sample.tokens(matchedWith: EmojiTokenizer())
         
         XCTAssert(tokens.count == 5, "Unexpected number of tokens [\(tokens.count)]")
