+++ Tests/FallbackTokenizerTests.swift
@@ -28,26 +28,26 @@ class FallbackTokenizerTests: XCTestCase {
     func testFallback() {
         
         let input = "1.2 34 abc catastrophe cat 0.5"
-        let tokens = input.tokens(matchedWith: NumberTokenizer.defaultTokenzier, "cat".literalTokenizer, CharacterSet.letters.anyTokenizer)
-
+        let tokens = input.tokens(matchedWith: NumberTokenizer.defaultTokenzier, "cat".literalTokenizer.anyTokenizer, CharacterSet.letters.anyTokenizer)
+        
         XCTAssert(tokens.count == 6, "Unexpected number of tokens [\(tokens.count)]")
         
-        //XCTAssert(tokens[0].tokenizer is NumberTokenizer)
+        XCTAssert(tokens[0] is NumberTokenizer.Token)
         XCTAssert(tokens[0].text == "1.2")
         
-        //XCTAssert(tokens[1].tokenizer is NumberTokenizer)
+        XCTAssert(tokens[1] is NumberTokenizer.Token)
         XCTAssert(tokens[1].text == "34")
         
-        //XCTAssert(tokens[2].tokenizer is CharacterSet)
+        XCTAssert(tokens[2] is CharacterSet.Token)
         XCTAssert(tokens[2].text == "abc")
         
-        //XCTAssert(tokens[3].tokenizer is CharacterSet)
+        XCTAssert(tokens[3] is CharacterSet.Token)
         XCTAssert(tokens[3].text == "catastrophe")
         
-        //XCTAssert(tokens[4].tokenizer is LiteralTokenizer)
+        XCTAssert(tokens[4] is LiteralTokenizer.Token)
         XCTAssert(tokens[4].text == "cat")
         
-        //XCTAssert(tokens[5].tokenizer is NumberTokenizer)
+        XCTAssert(tokens[5] is NumberTokenizer.Token)
         XCTAssert(tokens[5].text == "0.5")
     }
 }
