+++ Tests/MixedTokenTests.swift
@@ -23,78 +23,96 @@
 import XCTest
 import Mustard
 
-enum MixedTokenizer: TokenizerType, DefaultTokenizerType {
+final class MixedTokenizer: TokenizerType, DefaultTokenizerType {
     
-    case word
-    case number
-    case emoji
-    case none // 'none' case not strictly needed, and
-              // in this implementation will never be matched
-    init() {
-        self = .none
+    enum Mode {
+        case word
+        case number
+        case emoji
+        case none
     }
     
-    static let wordTokenizer = WordTokenizer()
-    static let numberTokenizer = NumberTokenizer()
-    static let emojiTokenizer = EmojiTokenizer()
+    var mode: Mode
+    
+    init() {
+        mode = .none
+    }
+    
+    static let wordTokenizer = WordTokenizer.defaultTokenzier
+    static let numberTokenizer = NumberTokenizer.defaultTokenzier
+    static let emojiTokenizer = EmojiTokenizer.defaultTokenzier
     
     func tokenCanTake(_ scalar: UnicodeScalar) -> Bool {
-        switch self {
+        switch mode {
         case .word: return MixedTokenizer.wordTokenizer.tokenCanTake(scalar)
         case .number: return MixedTokenizer.numberTokenizer.tokenCanTake(scalar)
         case .emoji: return MixedTokenizer.emojiTokenizer.tokenCanTake(scalar)
-        case .none:
-            return false
+        case .none: return false
         }
     }
     
     func tokenizerStartingWith(_ scalar: UnicodeScalar) -> AnyTokenizer? {
         
         if let _ = MixedTokenizer.wordTokenizer.tokenizerStartingWith(scalar) {
-            return MixedTokenizer.word.anyTokenizer
+            self.mode = .word
+            return self.anyTokenizer
         }
         else if let _ = MixedTokenizer.numberTokenizer.tokenizerStartingWith(scalar) {
-            return MixedTokenizer.number.anyTokenizer
+            self.mode = .number
+            return self.anyTokenizer
         }
         else if let _ = MixedTokenizer.emojiTokenizer.tokenizerStartingWith(scalar) {
-            return MixedTokenizer.emoji.anyTokenizer
+            self.mode = .emoji
+            return self.anyTokenizer
         }
         else {
+            self.mode = .none
             return nil
         }
     }
+    
+    struct MixedToken: TokenType {
+        let text: String
+        let range: Range<String.Index>
+        let mode: Mode
+    }
+    
+    func makeToken(text: String, range: Range<String.Index>) -> MixedToken {
+        return MixedToken(text: text, range: range, mode: mode)
+    }
+    
 }
 
 class MixedTokenTests: XCTestCase {
     
     func testMixedTokens() {
         
-        let tokens = "123👩‍👩‍👦‍👦Hello world👶again👶🏿45.67".tokens(matchedWith: MixedTokenizer.defaultTokenzier)
+        let tokens = "123👩‍👩‍👦‍👦Hello world👶again👶🏿45.67".tokens(matchedWith: MixedTokenizer())
         
         XCTAssert(tokens.count == 8, "Unexpected number of tokens [\(tokens.count)]")
         
-        //XCTAssert(tokens[0].tokenizer == .number)
+        XCTAssert(tokens[0].mode == .number)
         XCTAssert(tokens[0].text == "123")
         
-        //XCTAssert(tokens[1].tokenizer == .emoji)
+        XCTAssert(tokens[1].mode == .emoji)
         XCTAssert(tokens[1].text == "👩‍👩‍👦‍👦")
         
-        //XCTAssert(tokens[2].tokenizer == .word)
+        XCTAssert(tokens[2].mode == .word)
         XCTAssert(tokens[2].text == "Hello")
         
-        //XCTAssert(tokens[3].tokenizer == .word)
+        XCTAssert(tokens[3].mode == .word)
         XCTAssert(tokens[3].text == "world")
     
-        //XCTAssert(tokens[4].tokenizer == .emoji)
+        XCTAssert(tokens[4].mode == .emoji)
         XCTAssert(tokens[4].text == "👶")
         
-        //XCTAssert(tokens[5].tokenizer == .word)
+        XCTAssert(tokens[5].mode == .word)
         XCTAssert(tokens[5].text == "again")
         
-        //XCTAssert(tokens[6].tokenizer == .emoji)
+        XCTAssert(tokens[6].mode == .emoji)
         XCTAssert(tokens[6].text == "👶🏿")
         
-        //XCTAssert(tokens[7].tokenizer == .number)
+        XCTAssert(tokens[7].mode == .number)
         XCTAssert(tokens[7].text == "45.67")
         
     }
